requirements.txt
fastapi=0.67.0
aiofiles=0.7.0
python-multipart=0.0.5
orjson=3.6.0
psycopg2-binary=2.9.1
